2.0 Release
No previous issues were addressed rather this is an update for the sole purpose of increasing performance and reducing the amount of queries required. Name caching was implemented to a higher degree and a task installed to maintain the cache with the proper display information. Serveral core functions got minor adjustments but functionality has not changed.